Extract `toc/create_collection.rs`
